Create new container component to for single cards.
This allows us to send an ajax call for the coin price as the card is
rendered.

Because we are calling cryptocompare API for prices it made
sense to also change the cryptocompare URI constant to be only the base.
We can append paths and params as needed. Most of that will end up as
user configured state anyway.